+++ src/js/common/schemaform/reducers/index.js
@@ -2,13 +2,14 @@ import _ from 'lodash/fp';
 import { getDefaultFormState } from 'react-jsonschema-form/lib/utils';
 
 import {
-  setHiddenFields,
-  removeHiddenData,
-  updateRequiredFields,
   createFormPageList,
-  updateSchemaFromUiSchema,
-  replaceRefSchemas
 } from '../helpers';
+
+import {
+  updateSchemaAndData,
+  replaceRefSchemas,
+  updateItemsSchema
+} from '../formState';
 
 import { SET_DATA,
   SET_EDIT_MODE,
@@ -34,16 +35,8 @@ function recalculateSchemaAndData(initialState) {
 
       // Recalculate any required fields, based on the new data
       const page = state[pageKey];
-      let schema = updateRequiredFields(page.schema, page.uiSchema, formData);
 
-      // Update the schema with any fields that are now hidden because of the data change
-      schema = setHiddenFields(schema, page.uiSchema, formData);
-
-      // Update the schema with any general updates based on the new data
-      schema = updateSchemaFromUiSchema(schema, page.uiSchema, page.data, state);
-
-      // Remove any data that's now hidden in the schema
-      const data = removeHiddenData(schema, page.data);
+      const { data, schema } = updateSchemaAndData(page.schema, page.uiSchema, formData, page.data, state);
 
       if (page.data !== data || page.schema !== schema) {
         const newPage = _.assign(page, {
@@ -64,7 +57,8 @@ export default function createSchemaFormReducer(formConfig) {
   const firstPassInitialState = createFormPageList(formConfig)
     .reduce((state, page) => {
       const definitions = _.assign(formConfig.defaultDefinitions || {}, page.schema.definitions);
-      const schema = replaceRefSchemas(page.schema, definitions, page.pageKey);
+      let schema = replaceRefSchemas(page.schema, definitions, page.pageKey);
+      schema = updateItemsSchema(schema);
       const data = getDefaultFormState(schema, page.initialData, schema.definitions);
 
       return _.set(page.pageKey, {
