+++ test/common/schemaform/validation.unit.spec.js
@@ -163,13 +163,14 @@ describe('Schemaform validations', () => {
       const errors = {};
       const validator = sinon.spy();
       const schema = {
-        items: {
+        type: 'array',
+        items: [{
           properties: {
             field: {
 
             }
           }
-        }
+        }]
       };
       const uiSchema = {
         items: {
@@ -188,7 +189,11 @@ describe('Schemaform validations', () => {
 
       uiSchemaValidate(errors, uiSchema, schema, formData);
 
-      expect(validator.calledWith(errors[0].field, formData[0].field, formData, schema.items.properties.field, undefined)).to.be.true;
+      expect(validator.calledWith(errors[0].field,
+        formData[0].field,
+        formData,
+        schema.items[0].properties.field,
+        undefined)).to.be.true;
     });
     it('should skip validation when array is undefined', () => {
       const errors = {};
