+++ test/util/schemaform-utils.jsx
@@ -6,33 +6,27 @@ import React from 'react';
 import SchemaForm from '../../src/js/common/schemaform/SchemaForm';
 
 import {
-  setHiddenFields,
-  removeHiddenData,
-  updateRequiredFields,
-  updateSchemaFromUiSchema,
-  replaceRefSchemas
-} from '../../src/js/common/schemaform/helpers';
+  replaceRefSchemas,
+  updateSchemaAndData
+} from '../../src/js/common/schemaform/formState';
 
 export class DefinitionTester extends React.Component {
   constructor(props) {
     super(props);
     const { state, uiSchema } = props;
     const pageData = props.data;
-    const formData = _.merge(props.formData, pageData);
+    const formData = props.schema.type === 'array'
+      ? pageData || []
+      : _.merge(props.formData, pageData);
 
     const definitions = _.merge(props.definitions || {}, props.schema.definitions);
-    let schema = replaceRefSchemas(props.schema, definitions);
-    schema = updateRequiredFields(schema, uiSchema, formData);
-    // Update the schema with any fields that are now hidden because of the data change
-    schema = setHiddenFields(schema, uiSchema, formData);
-    // Update the schema with any general updates based on the new data
-    schema = updateSchemaFromUiSchema(schema, uiSchema, pageData, state);
-    // Remove any data that's now hidden in the schema
-    const newData = removeHiddenData(schema, pageData);
+    const schema = replaceRefSchemas(props.schema, definitions);
+
+    const { data: newData, schema: newSchema } = updateSchemaAndData(schema, uiSchema, formData, pageData, state);
 
     this.state = {
       data: newData,
-      schema,
+      schema: newSchema,
       uiSchema,
       formData
     };
@@ -41,14 +35,11 @@ export class DefinitionTester extends React.Component {
     // console.log('DefinitionTester -> handleChange -> data:', data);
     const state = this.props.state;
     const uiSchema = this.state.uiSchema;
-    const formData = _.merge(this.props.formData, data);
-    let schema = updateRequiredFields(this.state.schema, uiSchema, formData);
-    // Update the schema with any fields that are now hidden because of the data change
-    schema = setHiddenFields(schema, uiSchema, formData);
-    // Update the schema with any general updates based on the new data
-    schema = updateSchemaFromUiSchema(schema, uiSchema, data, state);
-    // Remove any data that's now hidden in the schema
-    const newData = removeHiddenData(schema, data);
+    const formData = this.state.schema.type === 'array'
+      ? data
+      : _.merge(this.props.formData, data);
+
+    const { data: newData, schema } = updateSchemaAndData(this.state.schema, uiSchema, formData, data, state);
 
     this.setState({
       data: newData,
